Update my code

+++ basic_math.py
@@ -67,8 +67,8 @@ def get_mean(number_list):
             47
     """
     mean = sum(number_list)/len(number_list)
-    return int(mean)
-
+    return mean
+    
 
 def get_median(number_list):
     """
